Move cache management methods out to separate mod
Why you made the change:

I did this so that they could be easily shared across different repositories.
Beyond that, I realized that the driver management, key generation, and fetcher
concerns are all technically dependencies of the cache management. Therefore, I
shifted the inclusion or extension of those respectively down to the
CacheManagement module. This also makes the interface for creating new
repositories nicer.
